carprice/utils/main_utils.py
import sys
import yaml
from carprice.exception import CarPriceException



def read_yaml_file(file_path: str) -> dict:
    try:
        with open(file_path, "rb") as yaml_file:
            return yaml.safe_load(yaml_file)

    except Exception as e:
        raise CarPriceException(e, sys) from e